added two possibilities for output value, float or integer in comment for degree_conversion.py

+++ PyPart1/degree_conversion.py
@@ -1,3 +1,4 @@
 celsius = int(input('Temperature in C? '))
-print(str(celsius * 1.8 + 32) + ' F')
+print(str(int(celsius * 1.8 + 32)) + ' F')
 #conversion to float or integer value?
+#print(str(int(celsius * 1.8 + 32)) + ' F')
